+++ ase/formula.py
@@ -1,9 +1,9 @@
-from math import gcd
 import re
-from typing import Dict, Tuple, List, Sequence, Union
+from functools import lru_cache
+from math import gcd
+from typing import Dict, List, Sequence, Tuple, Union
 
-from ase.data import chemical_symbols, atomic_numbers
-
+from ase.data import atomic_numbers, chemical_symbols
 
 # For type hints (A, A2, A+B):
 Tree = Union[str, Tuple['Tree', int], List['Tree']]  # type: ignore
@@ -28,7 +28,7 @@ class Formula:
             Only allow real chemical symbols.
         format: str
             Reorder according to *format*.  Must be one of hill, metal,
-            abc or reduce.
+            ab2, a2b, periodic or reduce.
 
         Examples
         --------
@@ -54,7 +54,8 @@ class Formula:
         """
         if format:
             assert _tree is None and _count is None
-            if format not in {'hill', 'metal', 'abc', 'reduce'}:
+            if format not in {'hill', 'metal', 'abc', 'reduce', 'ab2', 'a2b',
+                              'periodic'}:
                 raise ValueError(f'Illegal format: {format}')
             formula = Formula(formula).format(format)
         self._formula = formula
@@ -101,7 +102,7 @@ class Formula:
         return self.from_dict(dct), N
 
     def stoichiometry(self) -> Tuple['Formula', 'Formula', int]:
-        """Reduce to unique stoichiomerty using "chemical symbols" A, B, C, ...
+        """Reduce to unique stoichiometry using "chemical symbols" A, B, C, ...
 
         Examples
         --------
@@ -128,7 +129,10 @@ class Formula:
 
         * ``'hill'``: alphabetically ordered with C and H first
         * ``'metal'``: alphabetically ordered with metals first
-        * ``'abc'``: count ordered first then alphabetically ordered
+        * ``'ab2'``: count-ordered first then alphabetically ordered
+        * ``'abc'``: old name for ``'ab2'``
+        * ``'a2b'``: reverse count-ordered first then alphabetically ordered
+        * ``'periodic'``: periodic-table ordered: period first then group
         * ``'reduce'``: Reduce and keep order (ABBBC -> AB3C)
         * ``'latex'``: LaTeX representation
         * ``'html'``: HTML representation
@@ -171,9 +175,24 @@ class Formula:
             result += sorted(result2)
             return dict2str(dict(result))
 
-        if fmt == 'abc':
+        if fmt == 'abc' or fmt == 'ab2':
             _, f, N = self.stoichiometry()
             return dict2str({symb: n * N for symb, n in f._count.items()})
+
+        if fmt == 'a2b':
+            _, f, N = self.stoichiometry()
+            return dict2str({symb: -n * N
+                             for n, symb
+                             in sorted([(-n, symb) for symb, n
+                                        in f._count.items()])})
+
+        if fmt == 'periodic':
+            count = self.count()
+            order = periodic_table_order()
+            items = sorted(count.items(),
+                           key=lambda item: order.get(item[0], 0))
+            return ''.join(symb + (str(n) if n > 1 else '')
+                           for symb, n in items)
 
         if fmt == 'reduce':
             symbols = list(self)
@@ -196,7 +215,7 @@ class Formula:
             return self._tostr('<sub>', '</sub>')
 
         if fmt == 'rest':
-            return self._tostr(r'\ :sub`', r'`\ ')
+            return self._tostr(r'\ :sub:`', r'`\ ')
 
         if fmt == '':
             return self._formula
@@ -369,12 +388,22 @@ class Formula:
         return '+'.join(parts)
 
 
-def dict2str(dct):
+def dict2str(dct: Dict[str, int]) -> str:
+    """Convert symbol-to-number dict to str.
+
+    >>> dict2str({'A': 1, 'B': 2})
+    'AB2'
+    """
     return ''.join(symb + (str(n) if n > 1 else '')
                    for symb, n in dct.items())
 
 
-def parse(f: str):  # -> Tree
+def parse(f: str) -> Tree:
+    """Convert formula string to tree structure.
+
+    >>> parse('2A+BC2')
+    [('A', 2), (['B', ('C', 2)], 1)]
+    """
     if not f:
         return []
     parts = f.split('+')
@@ -386,6 +415,11 @@ def parse(f: str):  # -> Tree
 
 
 def parse2(f: str) -> Tree:
+    """Convert formula string to tree structure (no "+" symbols).
+
+    >>> parse('10(H2O)')
+    [(([('H', 2), 'O'], 1), 10)]
+    """
     units = []
     while f:
         unit: Union[str, Tuple[str, int], Tree]
@@ -421,6 +455,13 @@ def parse2(f: str) -> Tree:
 
 
 def strip_number(s: str) -> Tuple[int, str]:
+    """Strip leading nuimber.
+
+    >>> strip_number('10AB2')
+    (10, 'AB2')
+    >>> strip_number('AB2')
+    (1, 'AB2')
+    """
     m = re.match('[0-9]*', s)
     assert m is not None
     return int(m.group() or 1), s[m.end():]
@@ -428,6 +469,7 @@ def strip_number(s: str) -> Tuple[int, str]:
 
 def tree2str(tree: Tree,
              sub1: str, sub2: str) -> str:
+    """Helper function for html, latex and rest formats."""
     if isinstance(tree, str):
         return tree
     if isinstance(tree, tuple):
@@ -461,6 +503,18 @@ non_metals = ['H', 'He', 'B', 'C', 'N', 'O', 'F', 'Ne',
               'Ge', 'As', 'Se', 'Br', 'Kr',
               'Sb', 'Te', 'I', 'Xe',
               'Po', 'At', 'Rn']
+
+
+@lru_cache()
+def periodic_table_order() -> Dict[str, int]:
+    """Create dict for sorting after period first then row."""
+    return {symbol: n for n, symbol in enumerate(chemical_symbols[87:] +
+                                                 chemical_symbols[55:87] +
+                                                 chemical_symbols[37:55] +
+                                                 chemical_symbols[19:37] +
+                                                 chemical_symbols[11:19] +
+                                                 chemical_symbols[3:11] +
+                                                 chemical_symbols[1:3])}
 
 
 # Backwards compatibility: