+++ ase/test/test_formula.py
@@ -6,8 +6,17 @@ from ase.formula import Formula
 
 def test_formula_things():
     assert Formula('A3B2C2D').format('abc') == 'DB2C2A3'
+    assert Formula('A3B2C2D').format('ab2') == 'DB2C2A3'
+    assert Formula('A3B2C2D').format('a2b') == 'A3B2C2D'
     assert str(Formula('HHOOO', format='reduce')) == 'H2O3'
     assert Formula('HHOOOUO').format('reduce') == 'H2O3UO'
+
+
+@pytest.mark.parametrize('f',
+                         ['SiC', 'MoS2', 'GaAs', 'CO', 'NH3'])
+def test_periodic(f):
+    for fmt in ['ab2', 'a2b', 'periodic']:
+        assert Formula(f, format=fmt).format('periodic') == f
 
 
 def test_atoms_formula_things():